+++ R/listdial.R
@@ -1,4 +1,4 @@
-# Copyright (C) 2022 Hibiki AI Limited <info@hibiki-ai.com>
+# Copyright (C) 2022-2023 Hibiki AI Limited <info@hibiki-ai.com>
 #
 # This file is part of nanonext.
 #
@@ -24,10 +24,13 @@
 #' @param url [default 'inproc://nanonext'] a URL to dial, specifying the
 #'     transport and address as a character string e.g. 'inproc://anyvalue' or
 #'     'tcp://127.0.0.1:5555' (see \link{transports}).
-#' @param autostart [default TRUE] whether to start the dialer. Set to FALSE if
-#'     setting configuration options on the dialer as it is not generally
-#'     possible to change these once started. Supplying a non-logical value will
-#'     error.
+#' @param autostart [default TRUE] whether to start the dialer (by default
+#'     asynchronously). Set to NA to start synchronously - this is less resilient
+#'     if a connection is not immediately possible, but avoids subtle errors
+#'     from attempting to use the socket before an asynchronous dial has completed.
+#'     Set to FALSE if setting configuration options on the dialer as it is not
+#'     generally possible to change these once started. Supplying a non-logical
+#'     value will error.
 #'
 #' @return Invisibly, an integer exit code (zero on success). A new Dialer
 #'     (object of class 'nanoDialer' and 'nano') is created and bound to the