+++ R/socket.R
@@ -1,4 +1,4 @@
-# Copyright (C) 2022 Hibiki AI Limited <info@hibiki-ai.com>
+# Copyright (C) 2022-2023 Hibiki AI Limited <info@hibiki-ai.com>
 #
 # This file is part of nanonext.
 #
@@ -33,7 +33,10 @@
 #'     'tcp://127.0.0.1:5555' (see \link{transports}).
 #' @param autostart [default TRUE] whether to start the dialer/listener. Set to
 #'     FALSE if you wish to set configuration options on the dialer/listener as
-#'     it is not generally possible to change these once started.
+#'     it is not generally possible to change these once started. For dialers
+#'     only: set to NA to start synchronously - this is less resilient if a
+#'     connection is not immediately possible, but avoids subtle errors from
+#'     attempting to use the socket before an asynchronous dial has completed.
 #' @param raw [default FALSE] whether to open raw mode sockets. Note: not for
 #'     general use - do not enable unless you have a specific need, such as for
 #'     use with \code{\link{device}} (refer to NNG documentation).